Revert "make changes according to design descision"
This reverts commit ff05685a9f480831925c16059a693d1299be3edb.
