+++ frontend/src/components/LeftMenu/LeftMenu.css
@@ -1,97 +1,55 @@
+
 .left-menu {
-  display: flex;
-  justify-content: space-between;
-  flex-direction: column;
-  padding-left: 0.5rem;
-  padding-right: 1rem;
-  box-shadow: 0px 4px 6px rgba(81, 81, 81, 0.1);
-  border-right: 1px solid #EBEBEB;
-  min-width: max-content;
-}
+    display: flex;
+    justify-content: space-between;
+    flex-direction: column;
+    padding-left: 0.5rem;
+    padding-right: 1rem;
+    box-shadow: 0px 4px 6px rgba(81, 81, 81, 0.1); 
+    border-right: 1px solid #EBEBEB;
+    min-width: max-content;
+  }
+  .left-menu .MuiSvgIcon-root{
+    color:#667085;
+  }
+  .title .MuiTypography-root{
+    font-size: 11px;
+    font-weight: 400;
+    line-height: 38px;
+    color: #98A2B3;
+    margin-left: 8px;
+    margin-top: 8px;
+  }
+  
+  .left-menu .MuiListItemIcon-root {
+    min-width: 2rem; 
+  }
+  
+  
+  .search-item {
+    padding-top: 4px; 
+    padding-bottom: 4px; 
+  }
+  
+  .left-menu .MuiListItemText-root {
+    margin: 0; 
+  }
 
-.left-menu .MuiSvgIcon-root {
-  color: #667085;
-}
+  .menu-item .MuiTypography-root{
+    width: 100%; 
+    font-size: 13px;
+    font-weight: 400;
+    line-height: 24px;
+    color: #344054;
+    display: inline;
+    justify-content: center;
+    padding-right: 1.5rem;
+  }
 
-.title .MuiTypography-root {
-  font-size: 11px;
-  font-weight: 400;
-  line-height: 38px;
-  color: #98A2B3;
-  margin-left: 8px;
-  margin-top: 8px;
-}
+  .left-menu .MuiListItemButton-root{
+    border-radius: 6px;
+  }
 
-.left-menu .MuiListItemIcon-root {
-  min-width: 2rem;
-}
-
-
-.search-item {
-  padding-top: 4px;
-  padding-bottom: 4px;
-}
-
-.left-menu .MuiListItemText-root {
-  margin: 0;
-}
-
-.menu-item .MuiTypography-root {
-  width: 100%;
-  font-size: 13px;
-  font-weight: 400;
-  line-height: 24px;
-  color: #344054;
-  display: inline;
-  justify-content: center;
-  padding-right: 1.5rem;
-}
-
-.left-menu .MuiListItemButton-root {
-  border-radius: 6px;
-}
-
-.bottom-menu {
-  margin-bottom: 20px;
-}
-
-.user-info {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding-bottom: 1rem;
-  padding-left: 16px;
-}
-
-.user-details {
-  display: flex;
-  flex-direction: column;
-}
-
-.user-name {
-  font-size: 13px;
-  font-weight: 600;
-  line-height: 20px;
-  color: #344054;
-}
-
-.user-role {
-  margin: 0;
-  font-size: 13px;
-  font-weight: 400;
-  line-height: 20px;
-  color: #344054;
-}
-
-.dropdown-button {
-  background-color: transparent;
-  border: none;
-  cursor: pointer;
-  padding: 0.4em 0.4em;
-  margin: 1px
-}
-
-.user-details-container {
-  display: flex;
-}
+  .bottom-menu{
+    margin-bottom: 20px;
+  }
