+++ frontend/src/components/LeftMenu/LeftMenu.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { List, ListItemIcon, ListItemText, Divider, ListItemButton } from '@mui/material';
 import {
   DirectionsBusFilledOutlined as DirectionsBusIcon,
@@ -12,13 +12,7 @@ import {
   MarkChatUnreadOutlined as ChatIcon,
 } from '@mui/icons-material';
 import './LeftMenu.css';
-import Logo from '../Logo/Logo';
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../../services/authProvider';
-import Avatar from '../Avatar/Avatar';
-import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
-import KeyboardArrowDownOutlinedIcon from '@mui/icons-material/KeyboardArrowDownOutlined';
-import DropdownMenu from '../DropdownMenu/DropdownMenu';
 
 const menuItems = [
   { text: 'Home', icon: <HomeIcon />, route: '/' },
@@ -27,9 +21,9 @@ const menuItems = [
   { text: 'Hints', icon: <TipsIcon />, route: '/hint' },
   { text: 'Checklist', icon: <ChecklistIcon /> },
   { text: 'MAKE AN ANNOUNCEMENT', title: true },
-  { text: 'Popups', icon: <SmsIcon />, route: '/popup' },
-  { text: 'Banners', icon: <SportsIcon />, route: '/banner' },
-  { text: 'Helper Links', icon: <LinkIcon />, route: '/link' },
+  { text: 'Popups', icon: <SmsIcon />, route: '/popup'},
+  { text: 'Banners', icon: <SportsIcon /> , route: '/banner'},
+  { text: 'Helper Links', icon: <LinkIcon /> , route: '/link'},
   { text: 'GET FEEDBACK', title: true },
   { text: 'Feedback', icon: <ChatIcon /> },
   { text: 'Surveys', icon: <ListIcon /> },
@@ -42,16 +36,8 @@ function LeftMenu() {
     if (route) navigate(route);
   };
 
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-  const { userInfo } = useAuth();
-
-  const handleDropdownClick = () => {
-    setIsDropdownOpen(!isDropdownOpen);
-  };
-
   return (
     <div className="left-menu">
-      <Logo isSidebar={true} />
       <List>
         {menuItems.map((item, index) => (
           item.title ? (
@@ -76,18 +62,6 @@ function LeftMenu() {
           </ListItemIcon>
           <ListItemText primary="Support" />
         </ListItemButton>
-      </div>
-      <div className="user-info">
-        <div className='user-details-container'>
-          <Avatar src="/vendetta.png" alt="User" size="medium" />
-          <div className="user-details">
-            <div className="user-name">{userInfo.fullName}</div>
-            <div className="user-role">{userInfo.role}</div>
-          </div>
-        </div>
-        <button className="dropdown-button" onClick={handleDropdownClick}>
-          {isDropdownOpen ? <>< KeyboardArrowUpIcon /><DropdownMenu /></> : <KeyboardArrowDownOutlinedIcon />}
-        </button>
       </div>
     </div>
   );
