+++ frontend/src/scenes/dashboard/Dashboard.module.scss
@@ -2,7 +2,7 @@
     display: flex;
     flex-direction: column;
     width: 100%;
-    padding: 1.5rem 3rem;
+    padding: 3rem;
     gap:1.5rem;
     
     .top{
