+++ frontend/src/scenes/hints/HintPage.css
@@ -3,7 +3,7 @@
   grid-template-columns: 3fr 1fr;
   grid-gap: 20px;
   width: 100%;
-  padding: 3rem;
+  padding: 2% 3%;
 }
 
 .hint-page-title {
