+++ frontend/src/scenes/tours/ProductTourStyles.css
@@ -1,5 +1,5 @@
 .product-page-container {
-  padding: 3rem;
+  padding: 2% 3%;
 }
 
 .product-page-header {
