+++ frontend/src/templates/HomePageTemplate/HomePageTemplate.css
@@ -10,7 +10,3 @@
     display: flex;
     flex-grow: 1;
 }
-
-.content-container > div:last-child {
-    padding-top: 1.5rem;
-}
