+++ frontend/src/templates/HomePageTemplate/HomePageTemplate.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import LeftMenu from "../../components/LeftMenu/LeftMenu";
+import Header from "../../components/Header/Header";
 import "./HomePageTemplate.css";
 import { Outlet } from "react-router-dom";
 
@@ -8,6 +9,7 @@ const HomePageTemplate = () => {
 
   return (
     <div className="container">
+      <Header/>
       <div className="content-container">
         <LeftMenu className="sidebar"/>
         <Outlet />
